requirements.txt
langchain
ipykernel
python-dotenv 
langchain_community
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
lxml
langchain_google_genai
streamlit